requirements.txt
asgiref==3.2.3
astroid==2.3.3
certifi==2019.11.28
chardet==3.0.4
Django==3.0.3
django-crispy-forms==1.9.0
django-filter==2.2.0
gunicorn==20.0.4
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.0.0
psycopg2==2.8.4
pylint==2.4.4
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
whitenoise==5.0.1
wrapt==1.11.2
